extract useUserSettings hook from index
